[README]: Tell the truth about clang
It's not that it doesn't build at all, it does build for very, very, very, very simple grammars. (clang can't handle math expression grammars with +, -, *, /, '(' and ')' ).
But as extremely simple grammars does not need a complex parser to work, it ok to say that clang may not be as awesome as alphyn, yet.