[README]: Fix (again) the grammar type
The grammar is not BNF (it uses the C++ syntax, not a string), but it works exactly the same as BNF.

It would be possible to use alphyn to parse BNF and produce a grammar it can understand. The only problem with this is attributes... How would they be expressed ? Still, that would be overkill, so I need it **absolutly**. (It may even be possible to add an EBNF adaptator to alphyn, but the attribute issue remains and may even be worse).